# app_package/static/swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Population and Territory API",
    "version": "1.0.0",
    "description": "Get Population and Territory information for a given area and year"
  },
  "servers": [
    {
      "url": "http://10.32.1.47:5000/api/",
      "description": "main server"
    },
    {
      "url": "https://area-info.onrender.com/api/",
      "description": "just in case"
    },
    {
      "url": "http://127.0.0.1:5000/api/",
      "description": "local"
    }
  ],
  "paths": {
    "/regions/main_info": {
      "get": {
        "tags": [
          "Module 11"
        ],
        "summary": "Territory information (recent year): geometry, population, density",
        "parameters": [
          {
            "$ref": "#/components/parameters/territory_id"
          },
          {
            "$ref": "#/components/parameters/show_level"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/main_info_200"
                }
              }
            }
          }
        }
      }
    },
    "/regions/detailed_info": {
      "get": {
        "tags": [
          "Module 11"
        ],
        "summary": "Territory detailed information (all years)",
        "parameters": [
          {
            "$ref": "#/components/parameters/territory_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/detailed_info_200"
                }
              }
            }
          }
        }
      }
    },
    "/migrations/main_info": {
      "get": {
        "tags": [
          "Module 21"
        ],
        "summary": "Migration information (recent year): num in/out, factors",
        "description": "Inflow of <young / able to work / old> migrants; <br>  Outflow of <young / able to work / old> migrants; <br> Migration factors: avgemployers, avgsalary, ...",
        "parameters": [
          {
            "$ref": "#/components/parameters/territory_id"
          },
          {
            "$ref": "#/components/parameters/show_level"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/mig_main_info_200"
                }
              }
            }
          }
        }
      }
    },
    "/migrations/detailed_info": {
      "get": {
        "tags": [
          "Module 21"
        ],
        "summary": "Migration detailed information (all years): num in/out, types, factors",
        "description": "Inflow of <across russia / outside region / inside region /  international / interregional / all / other countries / CIS counries> migrants; <br>  Outflow of the same migrants; <br> Migration factors: avgemployers, avgsalary, ... <br><br> // В пределах России, Внешняя (для региона) миграция, Внутрирегиональная,  Международная, Межрегиональная, Миграция, всего,  С другими зарубежными странами, Со странами СНГ",
        "parameters": [
          {
            "$ref": "#/components/parameters/territory_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/mig_detailed_info_200"
                }
              }
            }
          }
        }
      }
    },
    "/regions/density_data_full": {
      "get": {
        "tags": [
          "returns geojson"
        ],
        "summary": "Population density by parent id; children with separate geometry",
        "parameters": [
          {
            "$ref": "#/components/parameters/parent_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/density_data_full_200"
                }
              }
            }
          }
        }
      }
    },
    "/regions/density_data": {
      "get": {
        "tags": [
          "returns geojson"
        ],
        "summary": "Population density",
        "parameters": [
          {
            "$ref": "#/components/parameters/parent_id"
          },
          {
            "$ref": "#/components/parameters/given_year"
          },
          {
            "$ref": "#/components/parameters/last_only"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/density_json_200"
                }
              }
            }
          }
        }
      }
    },
    "/regions/pyramid_data": {
      "get": {
        "tags": [
          "returns json"
        ],
        "summary": "Age-sex pyramid",
        "parameters": [
          {
            "$ref": "#/components/parameters/territory_id"
          },
          {
            "$ref": "#/components/parameters/given_year"
          },
          {
            "$ref": "#/components/parameters/n_age_groups"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/pyramid_json_200"
                }
              }
            }
          }
        }
      }
    },
    "/regions/migration_data": {
      "get": {
        "tags": [
          "returns json"
        ],
        "summary": "Migration estimate",
        "parameters": [
          {
            "$ref": "#/components/parameters/territory_id"
          },
          {
            "$ref": "#/components/parameters/given_year"
          },
          {
            "$ref": "#/components/parameters/n_age_groups"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/migration_json_200"
                }
              }
            }
          }
        }
      }
    },
    "/regions/area_needs": {
      "get": {
        "tags": [
          "returns json"
        ],
        "summary": "Area needs as a weighted sum of population needs",
        "parameters": [
          {
            "$ref": "#/components/parameters/territory_id"
          },
          {
            "$ref": "#/components/parameters/given_year"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/area_needs_json_200"
                }
              }
            }
          }
        }
      }
    },
    "/regions/pop_needs": {
      "get": {
        "tags": [
          "returns json"
        ],
        "summary": "Population needs by age (expert opinion). Estimated percent for each category",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/pop_needs_json_200"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "parent_id": {
        "name": "parent_id",
        "default": 34,
        "in": "query",
        "description": "Parent id of desired areas. *Example: 34*",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
      "territory_id": {
        "name": "territory_id",
        "default": 34,
        "in": "query",
        "description": "Territory id. *Example: 34*",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
      "last_only": {
        "name": "last_only",
        "in": "query",
        "description": "To show only last year info. *Example: True*",
        "required": false,
        "default": true,
        "schema": {
          "type": "boolean"
        }
      },
      "show_level": {
        "name": "show_level",
        "default": 2,
        "in": "query",
        "description": "Territory levels to show. *Example: 2*",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
      "okato_id": {
        "name": "okato_id",
        "in": "query",
        "description": "OKATO id of the area. *Example: 41206000000*",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "given_year": {
        "name": "given_year",
        "default": 2020,
        "in": "query",
        "description": "Data for a given year. *Example: 2020*",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
      "n_age_groups": {
        "name": "n_age_groups",
        "default": 5,
        "in": "query",
        "description": "Number of ages in an age interval. *Defualt: 5*",
        "required": false,
        "schema": {
          "type": "integer"
        }
      }
    },
    "schemas": {
      "404": {
        "title": "Sample",
        "type": "object",
        "properties": {
          "placeholder": {
            "type": "string",
            "description": "Placeholder description"
          }
        }
      },
      "pyramid_json_200": {
        "title": "Sample",
        "type": "object",
        "properties": {
          "columns": {
            "$ref": "#/components/schemas/columns"
          },
          "index": {
            "$ref": "#/components/schemas/index"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "number"
            },
            "example": [
              [
                10192,
                -10937
              ],
              [
                13393,
                -12231
              ],
              [
                3816,
                -1809
              ]
            ]
          }
        }
      },
      "migration_json_200": {
        "title": "Sample",
        "type": "object",
        "properties": {
          "columns": {
            "$ref": "#/components/schemas/columns"
          },
          "index": {
            "$ref": "#/components/schemas/index"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "number"
            },
            "example": [
              [
                256,
                259
              ],
              [
                -15,
                41
              ],
              [
                391,
                310
              ]
            ]
          }
        }
      },
      "density_json_200": {
        "title": "Sample",
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "2020": {
              "type": "number",
              "example": 1786.1
            },
            "territory_id": {
              "type": "integer"
            },
            "name": {
              "type": "string",
              "example": "Большеврудское"
            },
            "2020_dnst": {
              "type": "number",
              "example": 3.6
            },
            "binned": {
              "type": "string",
              "example": "0 -- 10"
            },
            "geometry": {
              "example": ""
            }
          }
        }
      },
      "density_data_full_200": {
        "title": "Sample",
        "type": "array",
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "2020": {
                "type": "number"
              },
              "territory_id": {
                "type": "integer"
              },
              "name": {
                "type": "string"
              },
              "2020_dnst": {
                "type": "number"
              },
              "2020_dnst_binned": {
                "type": "string"
              },
              "geometry_villages": {
                "type": "string"
              },
              "geometry": {
                "type": "string"
              }
            }
          }
        },
        "example": [
          {
            "2023": 50211,
            "territory_id": 10,
            "name": "Волосовский",
            "2023_dnst": 18.58,
            "2023_dnst_binned": "10 - 100",
            "geometry_villages": "",
            "geometry": ""
          },
          {
            "2023": 59877,
            "territory_id": 97,
            "name": "Киришский",
            "2023_dnst": 19.54,
            "2023_dnst_binned": "10 - 100",
            "geometry_villages": "",
            "geometry": ""
          }
        ]
      },
      "main_info_200": {
        "title": "Sample",
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "territory_id": {
              "type": "integer",
              "example": 1
            },
            "name": {
              "type": "string",
              "example": "Ленинградская область"
            },
            "geometry": {
              "example": ""
            },
            "density": {
              "type": "number",
              "example": 24.12
            },
            "pop_all": {
              "type": "number",
              "example": 2023767
            },
            "pop_younger": {
              "type": "number",
              "example": 303286
            },
            "pop_can_work": {
              "type": "number",
              "example": 1416751
            },
            "pop_older": {
              "type": "number",
              "example": 303730
            },
            "koeff_death": {
              "type": "number",
              "example": 0
            },
            "koeff_birth": {
              "type": "number",
              "example": 0
            },
            "koeff_migration": {
              "type": "number",
              "example": 0
            }
          }
        }
      },
      "detailed_info_200": {
        "title": "Sample",
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "pop_df": {
              "type": "object",
              "example": [
                "..."
              ]
            },
            "groups_df": {
              "type": "object",
              "example": [
                "..."
              ]
            },
            "dynamic_pop_df": {
              "type": "object",
              "example": [
                "..."
              ]
            },
            "soc_pyramid_df": {
              "type": "object",
              "example": [
                "..."
              ]
            },
            "values_df": {
              "type": "object",
              "example": [
                "..."
              ]
            }
          }
        }
      },
      "mig_main_info_200": {
        "title": "Sample",
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "territory_id": {
              "type": "integer",
              "example": 34
            },
            "name": {
              "type": "string",
              "example": "Всеволожский муниципальный район"
            },
            "geometry": {
              "example": ""
            },
            "younger_in": {
              "type": "number",
              "example": 13974
            },
            "work_in": {
              "type": "number",
              "example": 36966
            },
            "old_in": {
              "type": "number",
              "example": 4891
            },
            "younger_out": {
              "type": "number",
              "example": 6588
            },
            "work_out": {
              "type": "number",
              "example": 18678
            },
            "old_out": {
              "type": "number",
              "example": 2229
            },
            "avgemployers": {
              "type": "number",
              "example": 70096
            },
            "other": {
              "type": "string",
              "example": "..."
            }
          }
        }
      },
      "mig_detailed_info_200": {
        "title": "Sample",
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "year": {
              "type": "number",
              "example": 2019
            },
            "russia_in": {
              "type": "number",
              "example": 58967
            },
            "outside_region_in": {
              "type": "number",
              "example": 56890
            },
            "inside_region_in": {
              "type": "number",
              "example": 6466
            },
            "international_in": {
              "type": "number",
              "example": 4389
            },
            "interregional_in": {
              "type": "number",
              "example": 52501
            },
            "all_mig_in": {
              "type": "number",
              "example": 63356
            },
            "other_countries_in": {
              "type": "number",
              "example": 239
            },
            "cis_counries_in": {
              "type": "number",
              "example": 4150
            },
            "russia_out": {
              "type": "number",
              "example": 21423
            },
            "other": {
              "type": "string",
              "example": "..."
            }
          }
        }
      },
      "area_needs_json_200": {
        "title": "Sample",
        "type": "object",
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Здравоохранение",
              "Жилье",
              "Питание",
              "Транспорт",
              "Социальные связи",
              "Образование",
              "Досуг",
              "Волонтерство"
            ]
          },
          "index": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "example": [
              0
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "type": "number"
            },
            "example": [
              [
                0.1500906567,
                0.1523598427,
                0.116607779,
                0.1289441921,
                0.1566710245,
                0.1071761415,
                0.1229699438,
                0.0651804198
              ]
            ]
          }
        }
      },
      "pop_needs_json_200": {
        "title": "Sample",
        "type": "object",
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Возраст",
              "Здравоохранение",
              "Жилье",
              "Питание",
              "Транспорт",
              "Социальные связи",
              "Образование",
              "Досуг",
              "Волонтерство"
            ]
          },
          "index": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "example": [
              0,
              1,
              2,
              3,
              4,
              5,
              6,
              7,
              8
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "example": [
              [
                "0-6",
                90,
                60,
                80,
                40,
                70,
                50,
                60,
                10
              ],
              [
                "7-11",
                80,
                50,
                60,
                50,
                80,
                80,
                70,
                20
              ]
            ]
          }
        }
      },
      "columns": {
        "type": "array",
        "items": {
          "type": "string"
        },
        "minItems": 2,
        "maxItems": 2,
        "example": [
          "Женщины",
          "Мужчины"
        ]
      },
      "index": {
        "type": "array",
        "items": {
          "type": "string"
        },
        "example": [
          "0-4",
          "5-9",
          "10-14"
        ]
      }
    }
  }
}